flask/requirements.txt
beautifulsoup4==4.9.1
blinker==1.4
bs4==0.0.1
click==7.1.2
dask==2021.10.0
DateTime==4.3
Flask==1.1.2
Flask-Sitemap==0.3.0
Flask-Uploads==0.2.1
itsdangerous==1.1.0
Jinja2==2.11.3
lxml==4.9.1
MarkupSafe==1.1.1
numpy==1.21.6
pandas==1.0.5
pymongo==3.10.1
pysam==0.16.0.1
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.4
requests-oauthlib==1.3.0
six==1.15.0
soupsieve==2.0.1
tqdm==4.46.1
uuid==1.30
Werkzeug==0.16.1
zope.interface==5.1.0
